Revert "Axel comments to use templates"
This reverts commit 8800ef089b62cf173147b68adb2ee993b7e72980.
