+++ src/hotspot/share/classfile/classLoaderDataGraph.cpp
@@ -320,8 +320,61 @@ LockedClassesDo::~LockedClassesDo() {
   }
 }
 
+
+// Iterating over the CLDG needs to be locked because
+// unloading can remove entries concurrently soon.
+template <bool keep_alive = true>
+class ClassLoaderDataGraphIteratorBase : public StackObj {
+  ClassLoaderData* _next;
+  Thread*          _thread;
+  HandleMark       _hm;  // clean up handles when this is done.
+  NoSafepointVerifier _nsv; // No safepoints allowed in this scope
+                            // unless verifying at a safepoint.
+
+public:
+  ClassLoaderDataGraphIteratorBase() : _next(ClassLoaderDataGraph::_head), _thread(Thread::current()), _hm(_thread) {
+    if (keep_alive) {
+      assert_locked_or_safepoint(ClassLoaderDataGraph_lock);
+    } else {
+      assert_at_safepoint();
+    }
+  }
+
+  ClassLoaderData* get_next() {
+    ClassLoaderData* cld = _next;
+    // Skip already unloaded CLD for concurrent unloading.
+    while (cld != NULL && !cld->is_alive()) {
+      cld = cld->next();
+    }
+    if (cld != NULL) {
+      if (keep_alive) {
+        // Keep cld that is being returned alive.
+        Handle(_thread, cld->holder());
+      }
+      _next = cld->next();
+    } else {
+      _next = NULL;
+    }
+    return cld;
+  }
+};
+
 using ClassLoaderDataGraphIterator = ClassLoaderDataGraphIteratorBase<true /* keep_alive */>;
 using ClassLoaderDataGraphIteratorNoKeepAlive = ClassLoaderDataGraphIteratorBase<false /* keep_alive */>;
+
+void ClassLoaderDataGraph::loaded_cld_do(CLDClosure* cl) {
+  ClassLoaderDataGraphIterator iter;
+  while (ClassLoaderData* cld = iter.get_next()) {
+    cl->do_cld(cld);
+  }
+}
+
+void ClassLoaderDataGraph::loaded_cld_do_no_keepalive(CLDClosure* cl) {
+  ClassLoaderDataGraphIteratorNoKeepAlive iter;
+  while (ClassLoaderData* cld = iter.get_next()) {
+    cl->do_cld(cld);
+  }
+}
 
 // These functions assume that the caller has locked the ClassLoaderDataGraph_lock
 // if they are not calling the function from a safepoint.
