+++ src/hotspot/share/classfile/classLoaderStats.cpp
@@ -24,9 +24,8 @@
 
 #include "precompiled.hpp"
 #include "classfile/classLoaderData.inline.hpp"
-#include "classfile/classLoaderDataGraph.inline.hpp"
+#include "classfile/classLoaderDataGraph.hpp"
 #include "classfile/classLoaderStats.hpp"
-#include "classfile/javaClasses.inline.hpp"
 #include "memory/classLoaderMetaspace.hpp"
 #include "oops/objArrayKlass.hpp"
 #include "oops/oop.inline.hpp"
@@ -67,7 +66,7 @@ void ClassLoaderStatsClosure::do_cld(ClassLoaderData* cld) {
   }
 
   if (cl != NULL) {
-    cls->_parent = java_lang_ClassLoader::parent<AS_NO_KEEPALIVE>(cl);
+    cls->_parent = java_lang_ClassLoader::parent_no_keepalive(cl);
     addEmptyParents(cls->_parent);
   }
 
@@ -153,19 +152,19 @@ void ClassLoaderStatsClosure::addEmptyParents(oop cl) {
     ClassLoaderStats* cls = _stats->put_if_absent(cl, &added);
     if (added) {
       cls->_class_loader = cl;
-      cls->_parent = java_lang_ClassLoader::parent<AS_NO_KEEPALIVE>(cl);
+      cls->_parent = java_lang_ClassLoader::parent_no_keepalive(cl);
       _total_loaders++;
     }
     assert(cls->_class_loader == cl, "Sanity");
 
-    cl = java_lang_ClassLoader::parent<AS_NO_KEEPALIVE>(cl);
+    cl = java_lang_ClassLoader::parent_no_keepalive(cl);
   }
 }
 
 
 void ClassLoaderStatsVMOperation::doit() {
   ClassLoaderStatsClosure clsc (_out);
-  ClassLoaderDataGraph::loaded_cld_do<false>(&clsc);
+  ClassLoaderDataGraph::loaded_cld_do_no_keepalive(&clsc);
   clsc.print();
 }
