+++ src/hotspot/share/classfile/javaClasses.cpp
@@ -4664,6 +4664,16 @@ void java_lang_ClassLoader::serialize_offsets(SerializeClosure* f) {
 }
 #endif
 
+oop java_lang_ClassLoader::parent(oop loader) {
+  assert(is_instance(loader), "loader must be oop");
+  return loader->obj_field(_parent_offset);
+}
+
+oop java_lang_ClassLoader::parent_no_keepalive(oop loader) {
+  assert(is_instance(loader), "loader must be oop");
+  return loader->obj_field_access<AS_NO_KEEPALIVE>(_parent_offset);
+}
+
 // Returns the name field of this class loader.  If the name field has not
 // been set, null will be returned.
 oop java_lang_ClassLoader::name(oop loader) {
